fix: Replace int with long long int

+++ src/memory_allocator.h
@@ -10,14 +10,14 @@
 #include <stdlib.h>
 #include <stdbool.h>
 typedef struct memory_allocator {
-    void* (*malloc)(void*, process_t*, int);
-    void (*info)(void*, process_t*, int);
-    void (*use)(void*, process_t*, int);
-    void (*free)(void*, process_t*, int);
+    void* (*malloc)(void*, process_t*, long long int);
+    void (*info)(void*, process_t*, long long int);
+    void (*use)(void*, process_t*, long long int);
+    void (*free)(void*, process_t*, long long int);
     void (*load)(void*, process_t*);
-    int (*load_time_left)(void*, process_t*);
-    int (*require_allocation)(void*, process_t*);
-    int (*page_fault)(void*, process_t*);
+    long long int (*load_time_left)(void*, process_t*);
+    long long int (*require_allocation)(void*, process_t*);
+    long long int (*page_fault)(void*, process_t*);
     void* structure;
 } memory_allocator_t;
 #endif //SCHEDULER_MEMORY_ALLOCATOR_H
